+++ Annotation/RequestContent.php
@@ -15,29 +15,51 @@ use Doctrine\Common\Annotations\Annotation\Target;
 final class RequestContent
 {
 	/**
-	 * @Attribute(name="constraints", required=false, type="string")
+	 * @Attribute(name="constraints", type="string", required=false)
 	 * @var string $constraints
 	 */
 	protected $constraints;
 	
 	/**
-	 * @Attribute(name="can_be_empty", required=false, type="boolean")
+	 * @Attribute(name="can_be_empty", type="boolean", required=false)
 	 * @var boolean $canBeEmpty
 	 */
 	protected $canBeEmpty;
 	
 	/**
-	 * @Attribute(name="enable_validation", required=false, type="boolean")
+	 * @Attribute(name="enable_validation", type="boolean", required=false)
 	 * @var boolean $enableValidation
 	 */
 	protected $enableValidation;
 	
 	/**
-	 * @Attribute(name="validator_static_method", required=false, type="string")
+	 * @Attribute(name="validation_error_message", type="string", required=false)
+	 * @var string $validationErrorMessage
+	 */
+	protected $validationErrorMessage;
+	
+	/**
+	 * @Attribute(name="translation", type="array", required=false)
+	 * @var array $translation
+	 */
+	protected $translation;
+	
+	/**
+	 * @Attribute(name="validator_static_method", type="string", required=false)
 	 * @var string $validatorStaticMethod
 	 * @deprecated Use instead "constraints" property
 	 */
 	protected $validatorStaticMethod;
+	
+	/**
+	 * @var string $translationDomain
+	 */
+	private $translationDomain;
+	
+	/**
+	 * @var string $translationParameters
+	 */
+	private $translationParameters;
 	
 	/**
 	 * @param array $data
@@ -52,6 +74,10 @@ final class RequestContent
 		if ($this->constraints === null && $this->enableValidation === true) {
 			throw new \InvalidArgumentException('You must define a "constraints" attribute for each @RequestContent annotation while request validation is enabled.');
 		}
+		
+		$this->validationErrorMessage = isset($data['validation_error_message']) ? (string) $data['validation_error_message'] : 'response.bad_request';
+		$this->translationDomain = isset($data['translation']['domain']) ? $data['translation']['domain'] : 'OkaApiBundle';
+		$this->translationParameters = isset($data['translation']['parameters']) ? $data['translation']['parameters'] : [];
 	}
 	
 	/**
@@ -76,6 +102,30 @@ final class RequestContent
 	public function isEnableValidation()
 	{
 		return $this->enableValidation;
+	}
+	
+	/**
+	 * @return string
+	 */
+	public function getValidationErrorMessage()
+	{
+		return $this->validationErrorMessage;
+	}
+	
+	/**
+	 * @return string
+	 */
+	public function getTranslationDomain()
+	{
+		return $this->translationDomain;
+	}
+	
+	/**
+	 * @return array
+	 */
+	public function getTranslationParameters()
+	{
+		return $this->translationParameters;
 	}
 	
 	/**